Changed the move categorization logic to make Legendary moves just based on damage numbers.
Updated Profile to Trainer, and updated all relevant code for this change.

Moved the test level up to trainer, now main is strictly for generating pets.

Added a reroll move system with its own mechanics and specialized odds. An user's pet now has a reroll counter, which stores the available amount of rerolls so it does not get lost if it is not used.